PCHR-1585: Move unit test
